Combine 3 search algorithms into one scripts, modifies XgbRegressorSkOptRecipe to allows 2 more parameters
